All m,d,p data moved to index.js
